working on client bundle
