+++ src/Header/Left/left.js
@@ -1,11 +1,21 @@
 import React from 'react';
 
+import panafoldIcon from '../../../img/panafold.png';
+
 import styles from './styles';
 
-const Left = ({ icon, style, text }) => (
+const Left = ({ style }) => (
   <div style={style}>
-    <img src={icon} style={styles.logo} alt="logo" />
-    {text()}
+    <img
+      alt="panafold"
+      onClick={() => {
+        window.location = 'https://panafold.co/';
+      }}
+      src={panafoldIcon}
+      style={styles.logo}
+    />
+    <span style={styles.textImage}>image</span>
+    <span style={styles.textNgine}>ngine</span>
   </div>
 );
