+++ src/Header/Left/styles.js
@@ -3,6 +3,26 @@ const logo = {
   width: 'auto',
 };
 
+const txt = {
+  fontSize: '24px',
+  display: 'inline-block',
+  maxHeight: '24px',
+  marginBottom: '6px',
+};
+
+const textImage = {
+  ...txt,
+  color: 'darkgray',
+  marginLeft: '8px',
+};
+
+const textNgine = {
+  ...txt,
+  color: 'gray',
+};
+
 export default {
   logo,
+  textImage,
+  textNgine,
 };
